+++ src/mock/handlers/index.ts
@@ -1,0 +1,4 @@
+import { notificationHandlers } from './notifications'
+import { tagHandlers } from './tags'
+
+export const handlers = [...notificationHandlers, ...tagHandlers]
